requirements.txt
python-docx
spacy
sentence-transformers
qdrant-client
openai
streamlit
aiohttp
httpx
torch 
torchvision 
torchaudio
pyyaml
gitpython
diagrams
graphviz
boto3
ruamel.yaml
plotly